refactor: split 3 oversized query-keys files to comply with 300-line max
- billing-keys.ts (490 -> 245): extract subscription-keys.ts (280 lines)
  with subscriptionsKeys, subscriptionQueries, subscriptionStatusQuery,
  billingMutations, and callBillingEdgeFunction
- financial-keys.ts (420 -> 211): extract expense-keys.ts (236 lines)
  with expenseKeys, taxDocumentKeys, expenseQueries, financialTaxQueries,
  financialMutations, and Expense/CreateExpenseInput types
- payment-keys.ts (323 -> 224): extract payment-verification-keys.ts (123 lines)
  with paymentVerificationKeys and paymentStatusQueries
- Update all consumer imports in use-billing.ts, use-billing-mutations.ts,
  use-expense-mutations.ts, and use-payments.ts
- Update CLAUDE.md factory files list to reflect new file structure
